requirements.txt
attrs==21.2.0
Automat==20.2.0
cffi==1.15.0
constantly==15.1.0
cryptography==35.0.0
cssselect==1.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jeepney==0.7.1
jmespath==0.10.0
loguru==0.5.3
lxml==4.6.4
notify-py==0.3.3
notify2==0.3.1
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-crontab==2.6.0
python-dateutil==2.8.2
queuelib==1.6.2
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
tinydb==4.5.2
Twisted==21.7.0
typing_extensions==4.0.0
w3lib==1.22.0
zope.interface==5.4.0
